Factor out issue-search.tsx.
