+++ frontend/lib/tests/pages/issue-pages.test.tsx
@@ -2,12 +2,11 @@ import fs from 'fs';
 import path from 'path';
 import React from 'react';
 
-import { IssuesRoutes, getIssueLabel, getIssueAreaImagePath, groupByTwo, IssueAutocomplete, doesAreaMatchSearch } from '../../pages/issue-pages';
+import { IssuesRoutes, getIssueLabel, getIssueAreaImagePath, groupByTwo } from '../../pages/issue-pages';
 import Routes from '../../routes';
 import { AppTesterPal } from '../app-tester-pal';
 import { IssueAreaInput } from '../../queries/globalTypes';
 import { IssueAreaMutation_output } from '../../queries/IssueAreaMutation';
-import ReactTestingLibraryPal from '../rtl-pal';
 import { ISSUE_AREA_CHOICES } from '../../issues';
 
 
@@ -67,30 +66,4 @@ test('groupByTwo() works', () => {
   expect(groupByTwo([1])).toEqual([[1, null]]);
   expect(groupByTwo([1, 2])).toEqual([[1, 2]]);
   expect(groupByTwo([1, 2, 3])).toEqual([[1, 2], [3, null]]);
-});
-
-test('IssueAutocomplete works', () => {
-  const mockChange = jest.fn();
-  const pal = new ReactTestingLibraryPal(
-    <IssueAutocomplete inputValue="" onInputValueChange={mockChange} />
-  );
-
-  pal.fillFormFields([[/search/i, "mice"]]);
-  expect(mockChange.mock.calls).toHaveLength(1);
-  pal.rr.rerender(
-    <IssueAutocomplete inputValue="mice" onInputValueChange={mockChange} />
-  );
-  mockChange.mockClear();
-  pal.clickListItem(/mice/i);
-  expect(mockChange.mock.calls[1][0]).toBe('Entire home and hallways - Mice');
-
-  pal.rr.rerender(
-    <IssueAutocomplete inputValue="zzzzzzz" onInputValueChange={mockChange} />
-  );
-  expect(pal.rr.container.querySelector('li')).toBeNull();
-});
-
-test("doesAreaMatchSearch() works", () => {
-  expect(doesAreaMatchSearch('HOME', 'heat')).toBe(true);
-  expect(doesAreaMatchSearch('BATHROOMS', 'refrigerator')).toBe(false);
 });
